+++ httpie/cli/argparser.py
@@ -15,12 +15,11 @@ from httpie.cli.argtypes import (
     parse_format_options,
 )
 from httpie.cli.constants import (
-    DEFAULT_FORMAT_OPTIONS, HTTP_GET, HTTP_POST, OUTPUT_OPTIONS,
-    OUTPUT_OPTIONS_DEFAULT,
-    OUTPUT_OPTIONS_DEFAULT_STDOUT_REDIRECTED, OUT_RESP_BODY, PRETTY_MAP,
-    PRETTY_STDOUT_TTY_ONLY, SEPARATOR_CREDENTIALS, SEPARATOR_GROUP_ALL_ITEMS,
-    SEPARATOR_GROUP_DATA_ITEMS, URL_SCHEME_RE,
-    OUTPUT_OPTIONS_DEFAULT_OFFLINE,
+    HTTP_GET, HTTP_POST, OUTPUT_OPTIONS, OUTPUT_OPTIONS_DEFAULT,
+    OUTPUT_OPTIONS_DEFAULT_OFFLINE, OUTPUT_OPTIONS_DEFAULT_STDOUT_REDIRECTED,
+    OUT_RESP_BODY, PRETTY_MAP, PRETTY_STDOUT_TTY_ONLY, RequestType,
+    SEPARATOR_CREDENTIALS,
+    SEPARATOR_GROUP_ALL_ITEMS, SEPARATOR_GROUP_DATA_ITEMS, URL_SCHEME_RE,
 )
 from httpie.cli.exceptions import ParseError
 from httpie.cli.requestitems import RequestItems
@@ -84,6 +83,7 @@ class HTTPieArgumentParser(argparse.ArgumentParser):
         )
         # Arguments processing and environment setup.
         self._apply_no_options(no_options)
+        self._process_request_type()
         self._process_download_options()
         self._setup_standard_streams()
         self._process_output_options()
@@ -95,14 +95,31 @@ class HTTPieArgumentParser(argparse.ArgumentParser):
             self._body_from_file(self.env.stdin)
         self._process_url()
         self._process_auth()
+
+        if self.args.compress:
+            # TODO: allow --compress with --chunked / --multipart
+            if self.args.chunked:
+                self.error('cannot combine --compress and --chunked')
+            if self.args.multipart:
+                self.error('cannot combine --compress and --multipart')
+
         return self.args
+
+    def _process_request_type(self):
+        request_type = self.args.request_type
+        self.args.json = request_type is RequestType.JSON
+        self.args.multipart = request_type is RequestType.MULTIPART
+        self.args.form = request_type in {
+            RequestType.FORM,
+            RequestType.MULTIPART,
+        }
 
     def _process_url(self):
         if not URL_SCHEME_RE.match(self.args.url):
             if os.path.basename(self.env.program_name) == 'https':
                 scheme = 'https://'
             else:
-                scheme = self.args.default_scheme + "://"
+                scheme = self.args.default_scheme + '://'
 
             # See if we're using curl style shorthand for localhost (:3000/foo)
             shorthand = re.match(r'^:(?!:)(\d*)(/?.*)$', self.args.url)
@@ -276,7 +293,7 @@ class HTTPieArgumentParser(argparse.ArgumentParser):
                        'data (key=value) cannot be mixed. Pass '
                        '--ignore-stdin to let key/value take priority. '
                        'See https://httpie.org/doc#scripting for details.')
-        self.args.data = getattr(fd, 'buffer', fd).read()
+        self.args.data = getattr(fd, 'buffer', fd)
 
     def _guess_method(self):
         """Set `args.method` if not specified to either POST or GET
@@ -337,6 +354,7 @@ class HTTPieArgumentParser(argparse.ArgumentParser):
             self.args.data = request_items.data
             self.args.files = request_items.files
             self.args.params = request_items.params
+            self.args.multipart_data = request_items.multipart_data
 
         if self.args.files and not self.args.form:
             # `http url @/path/to/file`