+++ CHANGELOG.rst
@@ -8,14 +8,17 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
 `2.3.0-dev`_ (unreleased)
 -------------------------
+
 * Added support for multipart upload streaming (`#684`_).
+* Added support for body-from-file upload streaming (``http httpbin.org/post @file``).
+* Added ``--chunked`` to allow chunked transfer encoding.
+* Added ``--multipart`` to allow ``multipart/form-data`` encoding for non-file ``--form`` requests as well.
+* Added ``--boundary`` to allow a custom boundary string for ``multipart/form-data`` requests.
 * Added support for combining cookies specified on the CLI and in a session file (`#932`_).
 * Added out of the box SOCKS support with no extra installation (`#904`_).
 * Added ``--quiet, -q`` flag to enforce silent behaviour.
-* Added ``--multipart`` to allow ``multipart/form-data`` encoding for non-file ``--form`` requests as well.
-* Added ``--boundary`` to allow a custom boundary string for ``multipart/form-data`` requests.
-* Removed Tox testing entirely (`#943`_).
 * Fixed the handling of invalid ``expires`` dates in ``Set-Cookie`` headers (`#963`_).
+* Removed Tox testing entirely (`#943`_).
 
 
 `2.2.0`_ (2020-06-18)